+++ rectangle.cpp
@@ -18,9 +18,9 @@ Rectangle::Rectangle() {
 }
 
 /// @brief init constructor 
-/// @param w width in pixels
-/// @param h length in pixels
-Rectangle::Rectangle(int w, int l) {
+/// @param w width
+/// @param h length
+Rectangle::Rectangle(float w, float l) {
 
 	width = 0;
 	length = 0;
@@ -36,8 +36,6 @@ Rectangle::Rectangle(int w, int l) {
 		cout << "WARNING: Rectangle - constructor: length should be > 0" << endl;
 	else
 		length = l;
-
-
 
 }
 
@@ -83,8 +81,8 @@ bool Rectangle::operator==(const Rectangle& r)
 }
 
 /// @brief set width of the object
-/// @param w width in pixels
-void Rectangle::SetWidth(int w) {
+/// @param w width
+void Rectangle::SetWidth(float w) {
 
 	if (w < 0) {
 		cout << "WARNING: Rectangle - SetWidth: width should be > 0" << endl;
@@ -96,8 +94,8 @@ void Rectangle::SetWidth(int w) {
 }
 
 /// @brief set length of the object
-/// @param l length in pixels
-void Rectangle::SetLength(int l) {
+/// @param l length
+void Rectangle::SetLength(float l) {
 
 	if (l < 0) {
 		cout << "WARNING: Rectangle - SetLength: length should be > 0" << endl;
@@ -109,9 +107,9 @@ void Rectangle::SetLength(int l) {
 }
 
 /// @brief set width and length of the object
-/// @param w width in pixels
-/// @param l length in pixels
-void Rectangle::SetDim(int w, int l) {
+/// @param w width
+/// @param l length
+void Rectangle::SetDim(float w, float l) {
 
 	SetWidth(w);
 	SetLength(l);
@@ -120,7 +118,7 @@ void Rectangle::SetDim(int w, int l) {
 
 /// @brief get width of the object
 /// @return width of the object
-int Rectangle::GetWidth() {
+float Rectangle::GetWidth() {
 
 	return width;
 
@@ -128,7 +126,7 @@ int Rectangle::GetWidth() {
 
 /// @brief get length of the object
 /// @return length of the object
-int Rectangle::GetLength() {
+float Rectangle::GetLength() {
 
 	return length;
 
@@ -137,7 +135,7 @@ int Rectangle::GetLength() {
 /// @brief get width and length of the object returning them in the given variables
 /// @param given variable for width in pixels
 /// @param given variable for length in pixels
-void Rectangle::GetDim(int &w, int &l) {
+void Rectangle::GetDim(float&w, float&l) {
 
 	w = width;
 	l = length;
@@ -147,14 +145,14 @@ void Rectangle::GetDim(int &w, int &l) {
 
 /// @brief function to calculate the area of a rectangle
 /// @return area if the rectangle
-int Rectangle::GetArea()
+float Rectangle::GetArea()
 {
 	return width*length;
 }
 
 /// @brief function to calculate the perimeter of a rectangle
 /// @return perimeter if the rectangle
-int Rectangle::GetPerimeter()
+float Rectangle::GetPerimeter()
 {
 	return 2*(length + width);
 }
