+++ rectangle.h
@@ -15,8 +15,8 @@ using namespace std;
 class Rectangle
 {
 private:
-	int length;
-	int width;
+	float length;
+	float width;
 
 public:
 
@@ -24,7 +24,7 @@ public:
 	Rectangle();
 
 	// init constructor
-	Rectangle(int w, int l);
+	Rectangle(float w, float l);
 
 	//copy constructor
 	Rectangle(const Rectangle &r);
@@ -36,16 +36,16 @@ public:
 	Rectangle& operator=(const Rectangle &r);
 	bool operator==(const Rectangle& r);
 	
-	void SetDim(int w, int l);
-	void SetLength(int l);
-	void SetWidth(int w);
+	void SetDim(float w, float l);
+	void SetLength(float l);
+	void SetWidth(float w);
 	
-	void GetDim(int &w, int &l);
-	int GetLength();
-	int GetWidth();
+	void GetDim(float&w, float&l);
+	float GetLength();
+	float GetWidth();
 	
-	int GetArea();
-	int GetPerimeter();
+	float GetArea();
+	float GetPerimeter();
 };
 
 #endif
